requirements.txt
#requirements.txt

openai
crewai
fastapi
gradio
uvicorn
python-dotenv
langchain-openai